requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
pyyaml>=5.4.0 